Breaking out the library a little bit so that support does not have to be a mega file.
