+++ operations/alt-sentinel/get/followReference.js
@@ -1,7 +1,7 @@
 var hardLink = require('./../util/hardlink');
 var createHardlink = hardLink.create;
 var onValue = require('./onValue');
-var isExpired = require('./../util/support').isExpired;
+var isExpired = require('./../util/isExpired');
 
 function followReference(model, root, node, referenceContainer, reference, seed, outputFormat) {
