+++ operations/alt-sentinel/util/isExpired.js
@@ -1,0 +1,5 @@
+var now = require('../../../src/Falcor').now;
+module.exports = function isExpired(node) {
+    var $expires = node.$expires === undefined && -1 || node.$expires;
+    return $expires !== -1 && $expires !== 1 && ($expires === 0 || $expires < now());
+};
