+++ operations/alt-sentinel/util/permuteKey.js
@@ -1,0 +1,45 @@
+module.exports = function permuteKey(key, memo) {
+    if (memo.isArray) {
+        if (memo.loaded && memo.rangeOffset > memo.to) {
+            memo.arrOffset++;
+            memo.loaded = false;
+        }
+
+        var idx = memo.arrOffset;
+        if (idx === key.length) {
+            memo.done = true;
+            return '';
+        }
+
+        var el = key[memo.arrOffset];
+        var type = typeof el;
+        if (type === 'object') {
+            if (!memo.loaded) {
+                memo.from = el.from || 0;
+                memo.to = el.to || el.length && memo.from + el.length - 1 || 0;
+                memo.rangeOffset = memo.from;
+                memo.loaded = true;
+            }
+
+
+            return memo.rangeOffset++;
+        } else {
+            memo.arrOffset++;
+            return el;
+        }
+    } else {
+        if (!memo.loaded) {
+            memo.from = key.from || 0;
+            memo.to = key.to || key.length && memo.from + key.length - 1 || 0;
+            memo.rangeOffset = memo.from;
+            memo.loaded = true;
+        }
+        if (memo.rangeOffset > memo.to) {
+            memo.done = true;
+            return '';
+        }
+
+        return memo.rangeOffset++;
+    }
+};
+
